Refactored code and linted deps for upgrade to ESM
