+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "root": true,
+  "extends": [
+    "@nuxtjs"
+  ]
+}
