--- .eslintrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-  root: true,
-  parserOptions: {
-    parser: 'babel-eslint',
-    sourceType: 'module'
-  },
-  extends: [
-    '@nuxtjs'
-  ],
-  rules: {
-    "space-before-function-paren": ["error", "never"],
-  }
-}
