--- babel.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  presets: [
-    [
-      '@babel/preset-env', {
-        targets: {
-          esmodules: true
-        }
-      }
-    ]
-  ]
-}
