+++ components/charts/Chart.vue
