--- highcharts/contexts.js
@@ -1,2 +1,0 @@
-// @ts-nocheck
-export const hcModsContext = require.context('highcharts/modules', true, /\.js$/)
