--- highcharts/module.js
@@ -1,29 +1,0 @@
-// @ts-nocheck
-const { resolve } = require('path')
-
-module.exports = function (moduleOptions) {
-  const options = {
-    ...this.options['highcharts'],
-    ...moduleOptions
-  }
-
-  this.addTemplate({
-    ssr: false,
-    src: resolve(__dirname, 'contexts.js'),
-    fileName: 'nuxt-highcharts/contexts.js'
-  })
-
-  this.addTemplate({
-    src: resolve(__dirname, 'components.js'),
-    fileName: 'nuxt-highcharts/components.js'
-  })
-
-  this.addPlugin({
-    ssr: false,
-    src: resolve(__dirname, 'plugin.js'),
-    fileName: 'nuxt-highcharts/plugin.js',
-    options
-  })
-}
-
-module.exports.meta = require('../package.json')
