+++ nuxt.config.js
@@ -1,21 +1,22 @@
-/** @type {import('highcharts/highcharts').Options} */ 
-const setOptions = { 
+/** @type {import('highcharts/highcharts').Options} */
+const setOptions = {
   lang: { // <-- we correctly get intellisense
     // decimalPoint: '.' // this can be changed to anything, like ',' and it works
   }
 }
 
 /** @type {import('highcharts/highcharts').Options} */
-const chartOptions = { // <-- we also get intellisense correctly. 
-  credits: {
-    enabled: true // <-- typing "e" suggests "enabled?" ok
-  }
-}
+// const chartOptions = { // <-- we also get intellisense correctly.
+//   credits: {
+//     enabled: true // <-- typing "e" suggests "enabled?" ok
+//   }
+// }
 // Important note: as of 12/2020 it seems...
 // per the API docs, the "setOptions" get passed to the "Highcharts.setOptions" method
 // while the chartOptions would get passed to the *Highcharts.chart* method
 
-module.exports = {
+export default {
+  components: true,
   telemetry: false,
   env: {},
   ssr: true,
@@ -57,24 +58,14 @@ module.exports = {
    */
   modules: [
     'bootstrap-vue/nuxt',
-    '~/highcharts/module.js',
-    'nuxt-socket-io'
+    '~/lib/module.js'
   ],
-  io: {
-    sockets: [{
-      name: 'main',
-      url: process.env.NODE_ENV === 'production' 
-        ? process.env.IO_HOST_PROD
-        : process.env.IO_HOST_DEV
-
-    }]
-  },
   highcharts: {
     exporting: true,
     setOptions
     // mapChart: { // Also works
     //   mapName: 'myMapName',
-    //   mapData: {importedjsonData...} 
+    //   mapData: {importedjsonData...}
     // }
   },
   /*
@@ -84,10 +75,10 @@ module.exports = {
     /*
      ** You can extend webpack config here
      */
-    extend(config, ctx) {},
-    parallel: true,
-    cache: true,
-    hardSource: true
+    extend (config, ctx) {}
+    // parallel: true,
+    // cache: true,
+    // hardSource: true
   },
   globals: {
     loadingTimeout: 5000
