+++ package.json
@@ -1,19 +1,22 @@
 {
   "name": "nuxt-highcharts",
-  "version": "1.0.8",
+  "version": "2.0.0",
   "description": "Highcharts for Nuxt",
   "repository": "richardeschloss/nuxt-highcharts",
   "license": "MIT",
+  "type": "module",
   "contributors": [
     {
       "name": "Richard Schloss <richard.e.schloss@protonmail.com>"
     }
   ],
+  "imports": {
+    "#root/*": "./*"
+  },
   "files": [
-    "highcharts/components.js",
-    "highcharts/contexts.js",
-    "highcharts/module.js",
-    "highcharts/plugin.js"
+    "lib/components.js",
+    "lib/module.js",
+    "lib/plugin.js"
   ],
   "main": "highcharts/module.js",
   "scripts": {
@@ -22,54 +25,26 @@
     "start": "cross-env node server/index.js",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "test": "cross-env NODE_ENV=test npm run test:unit:cov && npm run test:e2e:cov",
-    "test:unit": "cross-env TEST=unit ava --config unit.config.js",
-    "test:unit:cov": "cross-env TEST=unit nyc ava --timeout=10m --config unit.config.js",
-    "test:unit:watch": "cross-env TEST=unit ava --config unit.config.js --color --watch",
-    "test:e2e": "cross-env TEST=e2e ava --config e2e.config.js",
-    "test:e2e:cov": "cross-env TEST=e2e nyc --no-clean ava --timeout=10m --config e2e.config.js",
-    "test:e2e:watch": "cross-env TEST=e2e ava --config e2e.config.js --color --watch"
+    "test": "ava",
+    "test:watch": "ava --color --watch",
+    "test:cov": "c8 ava --timeout=10m"
   },
   "dependencies": {
-    "highcharts": "^9.1.2"
+    "highcharts": "^9.3.0"
   },
   "devDependencies": {
-    "@ava/babel": "^1.0.1",
-    "@babel/preset-env": "^7.9.6",
-    "@babel/register": "latest",
-    "@nuxtjs/eslint-config": "^1.1.2",
-    "@nuxtjs/eslint-module": "^1.2.0",
-    "@vue/test-utils": "^1.0.0-beta.33",
-    "ava": "^3.8.2",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-module-resolver": "^4.0.0",
-    "bootstrap-vue": "^2.14.0",
+    "@babel/core": "^7.15.8",
+    "@nuxtjs/eslint-config": "^6.0.1",
+    "ava": "^3.15.0",
+    "bootstrap-vue": "^2.21.2",
     "browser-env": "^3.3.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-ignore-erb": "^0.1.1",
-    "eslint-plugin-jest": "^23.13.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-nuxt": "^0.5.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-unicorn": "^20.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "jsdom": "^16.2.2",
+    "c8": "^7.10.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-jest": "^25.2.2",
+    "eslint-plugin-unicorn": "^37.0.1",
+    "jsdom": "^16.7.0",
     "jsdom-global": "^3.0.2",
-    "node-fetch": "^2.6.1",
-    "nuxt": "^2.14.12",
-    "nuxt-socket-io": "^1.0.25",
-    "nuxt-test-utils": "latest",
-    "nyc": "^15.0.1",
-    "vue": "^2.6.12",
-    "vue-client-only": "^2.0.0",
-    "vue-meta": "^2.4.0",
-    "vue-no-ssr": "^1.1.1",
-    "vue-router": "^3.4.9",
-    "vue-template-compiler": "^2.6.12",
-    "vuex": "^3.6.0"
+    "nuxt": "^2.15.8"
   },
   "optionalDependencies": {
     "@highcharts/map-collection": "^1.1.3"
