+++ pages/index.vue
@@ -1,18 +1,27 @@
 <template>
   <div id="app">
-    <navbar />
-    <client-only>
-      <nuxt-child />
-    </client-only>
+    <Navbar :charts="charts" @chartSelected="(chart) => selectedChart = chart" />
+    <component :is="selectedChart.name" />
   </div>
 </template>
 
 <script>
-import Navbar from '@/components/Navbar.vue'
-
 export default {
-  components: {
-    Navbar
+  data () {
+    const charts = []
+    for (const c in this.$options.components) {
+      if (c.startsWith('Charts')) {
+        charts.push({
+          name: c,
+          displayName: c.replace('Charts', ''),
+          component: this.$options.components[c]
+        })
+      }
+    }
+    return {
+      selectedChart: '',
+      charts
+    }
   }
 }
 </script>
@@ -25,8 +34,8 @@ export default {
   text-align: center;
   color: #2c3e50;
   margin: 0 auto;
-  /* margin-top: 60px;*/
-  width: 70%; 
+  /* margin-top: 60px; */
+  width: 70%;
 }
 
 h1, h2 {
